requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-js-asset==1.2.2
djangorestframework==3.12.2
environ==1.0
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
